fix: build Rust test targets on Windows
`cargo clippy --all-targets --all-features -- -D warnings` and `cargo test`
fail on Windows/MSVC before any test runs:

- `Cred::credtype()` returns the bindgen-bound C enum type, which is `c_int`
  on MSVC but `c_uint` on GCC/Clang, so comparing it with `u32`
  `CredentialType::bits()` only type-checks on non-Windows targets.
- `std::path::PathBuf` is used only by a `#[cfg(unix)]` test.
- The Windows arm of `try_link_dir` needs no `return`.

CI runs on ubuntu-latest only, so these stayed invisible.

+++ src-tauri/src/core/device_sync/git_repo.rs
@@ -593,7 +593,10 @@ mod tests {
         )
         .unwrap();
 
-        assert_eq!(credential.credtype(), git2::CredentialType::SSH_KEY.bits());
+        assert_eq!(
+            i64::from(credential.credtype()),
+            i64::from(git2::CredentialType::SSH_KEY.bits())
+        );
     }
 
     #[test]
